databricks_apps/fraud-detection-app_2024_10_25-04_24/streamlit-fraud-app/requirements.txt
streamlit==1.38.0
streamlit_navigation_bar
openai
transformers
scikit-learn
seaborn
plotly
pandas
pyspark
python-dotenv
databricks-sql
openpyxl
json-repair




